requirements.txt
Kivy==2.2.0
kivymd==1.1.1
python-dotenv==1.0.0
requests~=2.31.0

pytest==7.3.2
